requirements.txt
azure-core
azure-storage
pandas
requests
lhafile